Refactor folder structure and add indexes
