+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/index.ts
@@ -1,4 +1,4 @@
 export * from './basic-elements';
-export * from './conditional-node/conditional-node.component';
+export * from './nodes';
 export * from './workflow.models';
 export * from './workflow.service';
