+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/conditional/conditional-node.component.html
