+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/cypher/cypher-node.component.html
