+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/cypher/cypher-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-restart-node',
+  selector: 'app-cypher-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './restart-node.component.html',
+  templateUrl: './cypher-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class RestartNodeComponent extends NodeDirective {}
+export class CypherNodeComponent extends NodeDirective {}
