+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/cypher/index.ts
@@ -1,0 +1,1 @@
+export * from './cypher-node.component';
