+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/internet/index.ts
@@ -1,0 +1,1 @@
+export * from './internet-node.component';
