+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/internet/internet-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-rephrase-node',
+  selector: 'app-internet-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './rephrase-node.component.html',
+  templateUrl: './internet-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class RephraseNodeComponent extends NodeDirective {}
+export class InternetNodeComponent extends NodeDirective {}
