+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/nucliadb/index.ts
@@ -1,0 +1,1 @@
+export * from './nucliadb-node.component';
