+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/nucliadb/nucliadb-node.component.html
