+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/nucliadb/nucliadb-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-cypher-node',
+  selector: 'app-nucliadb-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './cypher-node.component.html',
+  templateUrl: './nucliadb-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class CypherNodeComponent extends NodeDirective {}
+export class NucliaDBNodeComponent extends NodeDirective {}
