+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/rephrase/index.ts
@@ -1,0 +1,1 @@
+export * from './rephrase-node.component';
