+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/rephrase/rephrase-node.component.html
