+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/rephrase/rephrase-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-sql-node',
+  selector: 'app-rephrase-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './sql-node.component.html',
+  templateUrl: './rephrase-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SqlNodeComponent extends NodeDirective {}
+export class RephraseNodeComponent extends NodeDirective {}
