+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/restart/index.ts
@@ -1,0 +1,1 @@
+export * from './restart-node.component';
