+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/restart/restart-node.component.html
