+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/restart/restart-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-internet-node',
+  selector: 'app-restart-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './internet-node.component.html',
+  templateUrl: './restart-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class InternetNodeComponent extends NodeDirective {}
+export class RestartNodeComponent extends NodeDirective {}
