+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/sql/index.ts
@@ -1,0 +1,1 @@
+export * from './sql-node.component';
