+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/sql/sql-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-nucliadb-node',
+  selector: 'app-sql-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './nucliadb-node.component.html',
+  templateUrl: './sql-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class NucliaDBNodeComponent extends NodeDirective {}
+export class SqlNodeComponent extends NodeDirective {}
