+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/summarize/index.ts
@@ -1,0 +1,1 @@
+export * from './summarize-node.component';
