+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/summarize/summarize-node.component.html
