+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/summarize/summarize-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-validation-node',
+  selector: 'app-summarize-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './validation-node.component.html',
+  templateUrl: './summarize-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ValidationNodeComponent extends NodeDirective {}
+export class SummarizeNodeComponent extends NodeDirective {}
