+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/validation/index.ts
@@ -1,0 +1,1 @@
+export * from './validation-node.component';
