+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/validation/validation-node.component.html
