+++ apps/dashboard/src/app/retrieval-agent/agent-dashboard/workflow/nodes/validation/validation-node.component.ts
@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { AgentBoxComponent, NodeDirective } from '../basic-elements';
+import { AgentBoxComponent, NodeDirective } from '../../basic-elements';
 
 @Component({
-  selector: 'app-summarize-node',
+  selector: 'app-validation-node',
   imports: [CommonModule, AgentBoxComponent, TranslateModule],
-  templateUrl: './summarize-node.component.html',
+  templateUrl: './validation-node.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SummarizeNodeComponent extends NodeDirective {}
+export class ValidationNodeComponent extends NodeDirective {}
